frontend/requirements.txt
# Streamlit Frontend Requirements

# Dashboard Framework
streamlit==1.29.0

# API Client
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Image Processing
Pillow>=10.0.0
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
